# doc/openapi.yaml
openapi: 3.0.0
info:
  title: Playzone API
  version: '0.1'
servers:
  - url: 'http://localhost:8000'
paths:
...
  /api/projects:
    parameters: []
    get:
      summary: Récupérer des projets
      tags:
        - project
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: integer
                    minimum: 0
                  page:
                    type: integer
                    minimum: 0
                  size:
                    type: integer
                    minimum: 1
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      operationId: get-api-projects
      security:
        - ADMIN: []
        - USER: []
      parameters:
        - schema:
            type: integer
            minimum: 0
          in: query
          name: page
          description: index de la pagination
          required: true
        - schema:
            type: integer
            minimum: 1
          in: query
          name: size
          description: taille de la pagination
          required: true
        - schema:
            type: string
            enum:
              - OWNER
              - SHARED
              - PUBLIC
          in: query
          name: from
          description: 'pour n''afficher que les projets créés par l''utilisateur connecté, ou les projets partagés avec l''utilisateur connecté, ou les projets publics de tous les utilisateurs, sinon affiche tous les projets si ce champ est omis et que l''utilisateur connecté est un administrateur'
    post:
      summary: Créer un projet
      operationId: post-api-projects-projectId
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
              description: '/api/projects/{projectId}'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Created'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      security:
        - ADMIN: []
        - USER: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectFormCreate'
      tags:
        - project
...
  '/api/projects/{projectId}/schema':
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: projectId
        in: path
        description: id du projet
        required: true
    get:
      summary: Récupérer le schéma d'un projet
      tags:
        - schema
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchemaTable'
              examples:
                Example 1:
                  value:
                    - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                      table: DimNace
                      fact: false
                      headers:
                        - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                          name: SKNace
                          type: NUMBER
                          pk: true
                      coord:
                        x: 10
                        'y': 20
                      log:
                        createdAt: '2019-08-24T14:15:22.000Z'
                        createdBy: user@example.com
                        updatedAt: '2019-08-24T14:15:22.000Z'
                        updatedBy: user@example.com
                    - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                      table: FactEffectifs
                      fact: true
                      headers:
                        - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                          name: NaceFK
                          type: STRING
                          pk: false
                          fk:
                            table: 01BX5ZZKBKACTAV9WEVGEMMVRY
                            field: 01BX5ZZKBKACTAV9WEVGEMMVRY
                      coord:
                        x: 10
                        'y': 20
                      log:
                        createdAt: '2019-08-24T14:15:22.000Z'
                        createdBy: user@example.com
                        updatedAt: '2019-08-24T14:15:22.000Z'
                        updatedBy: user@example.com
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      operationId: get-api-projects-projectId-schema
      security:
        - ADMIN: []
        - USER: []
      parameters: []
    put:
      summary: Modifier le schéma d'un projet
      operationId: put-api-projects-projectId-schema
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      security:
        - ADMIN: []
        - USER: []
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SchemaTableForm'
            examples:
              Example 1:
                value:
                  - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                    fact: false
                    headers:
                      - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                        pk: true
                    coord:
                      x: 10
                      'y': 20
                  - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                    fact: true
                    headers:
                      - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                        pk: false
                        fk:
                          table: 01BX5ZZKBKACTAV9WEVGEMMVRY
                          field: 01BX5ZZKBKACTAV9WEVGEMMVRY
                    coord:
                      x: 10
                      'y': 20
        description: ''
      tags:
        - schema
...
  '/api/projects/{projectId}/schema/tables':
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: projectId
        in: path
        required: true
        description: id du projet
    get:
      summary: Récupérer la liste des id des tables du schéma d'un projet
      tags:
        - schema
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
      operationId: get-api-projects-projectId-schema-tables
      security:
        - ADMIN: []
        - USER: []
  '/api/projects/{projectId}/schema/tables-deleted':
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: projectId
        in: path
        required: true
        description: id du projet
    get:
      summary: Récupérer la liste des id des tables (supprimées) du schéma d'un projet
      tags:
        - schema
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
      operationId: get-api-projects-projectId-schema-tables-deleted
      security:
        - ADMIN: []
        - USER: []
  '/api/projects/{projectId}/schema/tables/{tableId}':
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: projectId
        in: path
        required: true
        description: id du projet
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: tableId
        in: path
        required: true
        description: id de la table
    post:
      summary: Ajouter une table au schéma d'un projet
      operationId: post-api-projects-projectId-schema-tables-tableId
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      security:
        - ADMIN: []
        - USER: []
      tags:
        - schema
    delete:
      summary: Retirer une table du schéma d'un projet
      operationId: delete-api-projects-projectId-schema-tables-tableId
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
      security:
        - ADMIN: []
        - USER: []
      tags:
        - schema
...
components:
  schemas:
...
    Project:
      title: Project
      type: object
      properties:
        id:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name:
          type: string
          example: Projet Emplois 2023
        description:
          type: string
          example: Statistiques de l'emploi en 2023
        isPublic:
          type: boolean
        log:
          $ref: '#/components/schemas/HistoryLog'
      required:
        - id
        - name
        - description
        - isPublic
        - log
    ProjectFormCreate:
      title: ProjectFormCreate
      type: object
      properties:
        name:
          type: string
          example: Projet Emplois 2023
        description:
          type: string
          example: Statistiques de l'emploi en 2023
      required:
        - name
        - description
...
    SchemaTable:
      title: SchemaTable
      type: object
      properties:
        id:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
          description: id de la table
        table:
          type: string
        fact:
          type: boolean
        headers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: id de la colonne
                example: 01BX5ZZKBKACTAV9WEVGEMMVRY
              name:
                type: string
              pk:
                type: boolean
                description: Si la colonne est une clé primaire
              fk:
                type: object
                nullable: true
                properties:
                  table:
                    type: string
                    description: id de la table référencée
                    example: 01BX5ZZKBKACTAV9WEVGEMMVRY
                  field:
                    type: string
                    description: id de la colonne référencée
                    example: 01BX5ZZKBKACTAV9WEVGEMMVRY
                required:
                  - table
                  - field
            required:
              - id
              - name
              - pk
        log:
          $ref: '#/components/schemas/HistoryLog'
        coord:
          type: object
          required:
            - x
            - 'y'
          properties:
            x:
              type: number
            'y':
              type: number
      required:
        - id
        - table
        - fact
        - headers
        - log
        - coord
    SchemaTableForm:
      title: SchemaTableForm
      type: object
      properties:
        id:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
          description: id de la table
        fact:
          type: boolean
        headers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: 01BX5ZZKBKACTAV9WEVGEMMVRY
                description: id de la colonne
              pk:
                type: boolean
                description: Si la colonne est une clé primaire
              fk:
                type: object
                nullable: true
                properties:
                  table:
                    type: string
                    description: id de la table référencée
                    example: 01BX5ZZKBKACTAV9WEVGEMMVRY
                  field:
                    type: string
                    description: id de la colonne référencée
                    example: 01BX5ZZKBKACTAV9WEVGEMMVRY
                required:
                  - table
                  - field
            required:
              - id
              - pk
        coord:
          type: object
          required:
            - x
            - 'y'
          properties:
            x:
              type: number
            'y':
              type: number
      required:
        - id
        - fact
        - headers
        - coord
...
    ErrorForm:
      title: ErrorForm
      type: object
      properties:
        status:
          type: integer
          example: 400
        path:
          type: string
          example: /api/endpoint
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: name
              message:
                type: string
                example: required
            required:
              - field
              - message
      required:
        - status
        - path
        - errors
...
    ErrorUnauthorized:
      title: ErrorUnauthorized
      type: object
      properties:
        status:
          type: integer
          example: 401
        path:
          type: string
          example: /api/endpoint
        error:
          type: string
          example: Unauthorized
      required:
        - status
        - path
        - error
    ErrorForbidden:
      title: ErrorForbidden
      type: object
      properties:
        status:
          type: integer
          example: 403
        path:
          type: string
          example: /api/endpoint
        error:
          type: string
          example: Forbidden
      required:
        - status
        - path
        - error
    Created:
      title: Created
      type: object
      properties:
        id:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
      required:
        - id
    HistoryLog:
      title: HistoryLog
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: email
        updatedBy:
          type: string
          format: email
      required:
        - createdAt
        - updatedAt
        - createdBy
        - updatedBy
  securitySchemes:
    ADMIN:
      type: http
      scheme: bearer
      description: ''
    USER:
      type: http
      scheme: bearer
...